requirements.txt
streamlit
streamlit_chat
langchain
sentence_transformers
openai
chromadb
tiktoken
docx
docx2txt
pypdf
python-docx